grpc clients and etc

+++ internal/services/parcels_manager_service/internal/store/parcel/store.go
@@ -9,6 +9,6 @@ type Store interface {
 	Delete(model.TrackNumber) error
 	GetInfo(model.TrackNumber) (model.Parcel, error)
 	AddCheckpoint(model.TrackNumber, model.Checkpoint) error
-	GetCheckpoints(trackNum model.TrackNumber, page int, pageSize int) ([]*model.Checkpoint, error)
+	GetCheckpoints(trackNum model.TrackNumber, page uint64, pageSize uint64) ([]*model.Checkpoint, error)
 	Close()
 }
